[ALPS05860938] eint: update data type of wake_mask
update data type of wake_mask

MTK-Commit-Id: 3c4608ab230278c32e4c0e125fc3ba171122a735

CR-Id: ALPS05860938
Feature: EINT

+++ drivers/pinctrl/mediatek/mtk-eint.h
@@ -58,8 +58,8 @@ struct mtk_eint_instance {
 	void __iomem *base;
 	unsigned int number;
 	u8 pin_list[MAX_PIN];
-	u8 *wake_mask;
-	u8 *cur_mask;
+	unsigned int *wake_mask;
+	unsigned int *cur_mask;
 };
 
 struct mtk_eint;
